+++ include/Camera.h
@@ -23,7 +23,7 @@ class Camera {
     static void handleControllerAxis(float value, KeyboardButton axis);
 
     //! \fixme The Y axis seems to be the source of all evil?
-    static void setPosition(glm::vec3 p) { pos = glm::vec3(p.x, -p.y, p.z); }
+    static void setPosition(glm::vec3 p) { pos = glm::vec3(p.x, -p.y, p.z); dirty = true; }
     static glm::vec3 getPosition() { return glm::vec3(pos.x, -pos.y, pos.z); }
 
     static glm::vec2 getRotation() { return rot; }